Increase possible paper image size to 140px high
With the aspect ratio fix, some images become quite small. This extends
the allowed image size to be 140px in height. Manual inspection shows
this does not cause any paper images to overflow their container.

A more appropriate fix would be to use a flexbox container, however I
wanted to to make as few code changes as possible.